### 2021-10-07 #90 @whedon
[PRE REVIEW]: ProbabilityBoundsAnalysis.jl: theory and algorithms
**Submitting author:** @AnderGray (<a href="http://orcid.org/0000-0002-1585-0900">Ander Gray</a>)
**Repository:** <a href="https://github.com/AnderGray/ProbabilityBoundsAnalysis.jl" target ="_blank">https://github.com/AnderGray/ProbabilityBoundsAnalysis.jl</a>
**Version:**
