### 2021-10-22 #96 @whedon
[REVIEW]: ProbabilityBoundsAnalysis.jl: theory and algorithms
**Submitting author:** @AnderGray (<a href="http://orcid.org/0000-0002-1585-0900">Ander Gray</a>)
**Repository:** https://github.com/AnderGray/ProbabilityBoundsAnalysis.jl
**Branch with paper.md** (empty if default branch):
**Version:**
