+++ paper/header.tex
@@ -11,7 +11,7 @@
 \keywords{Julia, Probability, Uncertainty, Interval Arithmetic, Probabilistic Arithmetic}
 
 \hypersetup{
-pdftitle = {ProbabilityBoundsAnalysis.jl:\\ an Arithmetic of probabilities and sets of probabilities},
+pdftitle = {ProbabilityBoundsAnalysis.jl: Arithmetic with sets of distributions},
 pdfsubject = {JuliaCon 2019 Proceedings},
 pdfauthor = {Ander Gray, Scott Ferson, Edoardo Patelli},
 pdfkeywords = {Julia, Uncertainty, Interval Arithmetic, Probability Bounds Analysis, Probability Boxes},
