### 2021-11-08 #96 @AnderGray
@dpsanders @baggepinnen, not sure why it's not finding the paper. I changed the title in the header.tex, could this perhaps have led to some issues? Was it ok to change the title?

Otherwise, a pdf of the paper is uploaded to the branch
