### 2021-11-14 #96 @carstenbauer
Perhaps the issue here is that the branch information ("JuliaConPaper") is already part of the repo URL (see OP). @vchuravy, can you drop that part and make the repo here simply point to https://github.com/AnderGray/ProbabilityBoundsAnalysis.jl? (I think I can't do it, right?)

@AnderGray In any case, it is better to not have a `paper.pdf` in the `paper/` folder of the `JuliaConPaper` branch. On the one hand, to not confuse whedon, and on the other hand since the paper isn't published yet.
