### 2021-12-14 #96 @baggepinnen
I have started my review, but I've been reading the pdf that was deleted from the  https://github.com/AnderGray/ProbabilityBoundsAnalysis.jl repo, I couldn't find any other pdf?
