### 2022-02-06 #96 @baggepinnen
I have almost completed the review. There are currently 4 uncecked boxes,
- The automated tests appear to cover a very minor part of the library only so I can't really check this box, most of the library is not automatically tested? The very first thing I tried also failed, so I take that as the testing needs to improve.
- I'm not sure where to find the page limit so I can't verify it.
- There are no DOIs in the reference list
- I can't find any community guidlines in the repo

The thing I tested was the following

I’m surprised IntervalArithmetic.jl, which is used by the reviewd library, does not error here but rather picks one of the branches
```julia
julia> a = 1..3
[1, 3]

julia> foo(x) = x > 2 ? 10 : 20
foo (generic function with 1 method)

julia> foo(a)
20
```
If I used this to represent a probability distribution, the result would be incorrect since the true interval is 10..20

PBA errors for the example above (assuming I’m using it right)
```julia
julia> a = pbox(1,3)
Pbox: 	  ~  ( range=[1.0, 3.0], mean=[1.0, 3.0], var=[0.0, 1.0])

julia> foo(a)
ERROR: promotion of types Int64 and pbox failed to change any arguments
```
But the error message does not indicate that I did something wrong.
