### 2022-02-07 #96 @dpsanders
@carstenbauer: Apologies for the delay.

The links to the JuliaCon conflict of interest and code of conduct seem to be broken.
