### 2022-02-07 AnderGray/ProbabilityBoundsAnalysis.jl#31 @dpsanders
Make `shape` a type
It would be more "Julian" to make the shape of a pbox into a type, e.g. `Normal()`, `Uniform()`, etc.

https://github.com/AnderGray/ProbabilityBoundsAnalysis.jl/blob/652700c7d2e1853f189b463d0786fd2f3c467f16/src/pbox/pbox.jl#L62

Part of the review for the JuliaCon proceedings: https://github.com/JuliaCon/proceedings-review/issues/96
