### 2022-02-07 AnderGray/ProbabilityBoundsAnalysis.jl#32 @dpsanders
Clashes with `Plots.jl`
My default graphing library is `Plots.jl`, but PBA clashes with it:

```jl
julia> a = uniform(0, 1)
Pbox: 	  ~ uniform ( range=[0.0, 1.0], mean=0.5, var=0.083333)

julia> using Plots

julia> plot(a)
WARNING: both Plots and ProbabilityBoundsAnalysis export "plot"; uses of it in module Main must be qualified
```

I'm not sure what is actionable here, but it should at least be mentioned in the README and docs.

Part of the review for the JuliaCon proceedings: https://github.com/JuliaCon/proceedings-review/issues/96
