### 2022-02-07 AnderGray/ProbabilityBoundsAnalysis.jl#32 @dpsanders
It would be great to make [plot recipes](https://docs.juliaplots.org/latest/recipes/) for Plots.jl. You could maybe detect which plotting package is available and define code accordingly.

(But all that is necessary for the review is to add a note about this issue.)
