### 2022-02-07 #96 @dpsanders
👋 @AnderGray: It's not clear to me why Eduardo Patelli is an author of the paper, since he does not seem to be an author of the package, according to the docs.
