### 2022-02-17 d65f2e7
removes pyplot and pycall dependence
