+++ src/pbox/copulas.jl
@@ -43,10 +43,6 @@ include("NormalDistribution.jl")
 #global bOt = 0.001;              # smallest quamtile to use if left tail is unbounded
 #global tOp = 0.999;              # largest quamtile to use if right tail is unbounded
 
-pyimport_conda("mpl_toolkits.mplot3d", "mpl_toolkits")
-art3d = PyObject(PyPlot.art3D)
-mpl = pyimport("matplotlib")
-using3D()
 
 abstract type AbstractCopula <: Real end
 abstract type AbstractJoint <: Real end
@@ -119,7 +115,7 @@ end
 
 (obj::copula)(X :: StepRangeLen, Y :: Real, useInterp = false)  = [obj(i,Y,useInterp) for i in X]
 
-function (obj::copula)(X :: Interval, Y::Interval, useInterp = false) 
+function (obj::copula)(X :: Interval, Y::Interval, useInterp = false)
 
     cdfLo = obj(X.lo,Y.lo, useInterp).lo
     cdfHi = obj(X.hi,Y.hi, useInterp).hi
@@ -128,12 +124,12 @@ function (obj::copula)(X :: Interval, Y::Interval, useInterp = false)
 
 end
 
-function (obj::copula)(X :: Interval, Y::Real, useInterp = false) 
+function (obj::copula)(X :: Interval, Y::Real, useInterp = false)
     Y = interval(Y);
     return obj(X,Y,useInterp)
 end
 
-function (obj::copula)(X :: Real, Y::Interval, useInterp = false) 
+function (obj::copula)(X :: Real, Y::Interval, useInterp = false)
     X = interval(X);
     return obj(X,Y,useInterp)
 end
@@ -157,7 +153,7 @@ function cdfU(C :: copula, X, Y)
 
     #w = [opp(x,y) for x in X, y in Y]
     #m = [perf(x,y) for x in X, y in Y]
-    
+
     #cdfLo = max.(C.cdfU[xIndexLower, yIndexLower], w)
     #cdfHi = min.(C.cdfU[xIndexUpper, yIndexUpper], m)
 
@@ -177,10 +173,10 @@ function cdfD(C :: copula, X, Y)
 
     xIndexUpper = Int.(ceil.(X .* (nX-1)))  .+ 1
     yIndexUpper = Int.(ceil.(Y .* (nY-1)))  .+ 1
-    
+
     #w = [opp(x,y) for x in X, y in Y]
     #m = [perf(x,y) for x in X, y in Y]
-    
+
     #cdfLo = max.(C.cdfD[xIndexLower, yIndexLower], w)
     #cdfHi = min.(C.cdfD[xIndexUpper, yIndexUpper], m)
 
@@ -256,7 +252,7 @@ function conditionalX(C :: AbstractCopula, xVal :: Real)
     Nsx, Nsy = size(C.cdfU)
 
     n = ProbabilityBoundsAnalysis.steps
-    
+
     yGridU = ProbabilityBoundsAnalysis.ii();
     yGridD = ProbabilityBoundsAnalysis.jj();
 
@@ -266,7 +262,7 @@ function conditionalX(C :: AbstractCopula, xVal :: Real)
     zD = (right.(C(xVal + diff/2, yGridD) - right.(C(xVal - diff/2, yGridD))))/diff
 
     ###
-    #   zU are the cdf values, and yGridU/D are the physical values, which is a uniform grid. 
+    #   zU are the cdf values, and yGridU/D are the physical values, which is a uniform grid.
[…]
