+++ src/pbox/plot_recipes.jl
@@ -96,6 +96,420 @@ function prepFillBounds(x)
     return Xs, Ylb[:], Yub[:]
 end
 
+###
+#       Copulas and bivariate p-box plots
+###
+
+function plotStep(cop :: copula; name = missing, pn = 50, fontsize = 18, alpha = 0.3)
+
+    AU = cop.cdfU
+    AD = cop.cdfD
+    m = size(AU)[1];
+    if m < pn; ppn = m; else ppn = pn; end      #pn is the plot "sub-sample", we can't plot all elements
+
+    x = y = range(0, stop = 1,length=m)
+    #xgrid = repeat(x',ppn,1)
+    #ygrid = repeat(y,1,ppn)
+
+    nm = round(m/ppn);
+
+    x = x[1:Int(nm):end]
+    oneIn = x[end] == 1
+
+    if !oneIn; x = [x; 1]; end                             # We always want to include 1 in the plot
+
+    zU = AU[1:Int(nm):end,1:Int(nm):end]
+    zD = AD[1:Int(nm):end,1:Int(nm):end]
+
+    if !oneIn;
+
+        zU = [zU AU[end, 1:Int(nm):end]]
+        zU = [zU' [AU[1:Int(nm):end, end]; 1]]
+
+        zD = [zD AD[end, 1:Int(nm):end]]
+        zD = [zD' [AD[1:Int(nm):end, end]; 1]]
+    end
+
+    if ismissing(name); fig = figure(figsize=(10,10)) else fig = figure(name,figsize=(10,10));end
+    ax = fig.add_subplot(1,1,1,projection="3d")
+
+    Ufaces = prepPolysU(zU,x);
+    Dfaces = prepPolysD(zD,x);
+
+    p3Ufaces = PyObject(art3d.Poly3DCollection(Ufaces, alpha=alpha))
+    p3Dfaces = PyObject(art3d.Poly3DCollection(Dfaces, alpha=1))
+
+    face_colorU = [1, 0, 0]
+    face_colorD = [0, 0, 1]
+    edge_color = [0.4, 0.4, 0.4]
+
+    pycall(p3Ufaces.set_facecolor, PyAny, face_colorU)
+    pycall(p3Ufaces.set_edgecolor, PyAny, edge_color)
+
+    pycall(p3Dfaces.set_facecolor, PyAny, face_colorD)
+    pycall(p3Dfaces.set_edgecolor, PyAny, edge_color)
+
+    pycall(ax.add_collection3d, PyAny, p3Dfaces)
+    pycall(ax.add_collection3d, PyAny, p3Ufaces)
+
+
+end
+
+function prepPolysU(us, x)
+
+    ppn = length(x)
+
+    p1 = [x[1]; x[1]; us[2, 2]]
+    p2 = [x[1]; x[2]; us[2, 2]]
+    p3 = [x[2]; x[2]; us[2, 2]]
+    p4 = [x[2]; x[1]; us[2, 2]]
+    tops = [[tuple(p1...); tuple(p2...); tuple(p3...); tuple(p4...)]]
+
+    for i = 1:1:ppn-1
+        for j = 1:1:ppn-1
+            p1 = [x[i]; x[j]; us[i+1,j+1]]
+            p2 = [x[i]; x[j+1]; us[i+1, j+1]]
+            p3 = [x[i+1]; x[j+1]; us[i+1, j+1]]
+            p4 = [x[i+1]; x[j]; us[i+1, j+1]]
+            this = [[tuple(p1...); tuple(p2...); tuple(p3...); tuple(p4...)]]
+            tops = [tops; this]
+        end
+    end
+
+    tops = tops[2:end]
+
+    p1 = [x[1]; x[1]; us[1, 1]]
+    p2 = [x[1]; x[1]; us[2, 2]]
[…]
