+++ .travis.yml
@@ -1,35 +1,0 @@
-language: julia
-
-os: 
-  - linux
-
-julia: 
-  - nightly
-  - 1.4
-  - 1.5
-  - 1.6
-
-
-notifications:
-  email: false
-
-before_install: 
-  - "julia -e 'ENV[\"PYTHON\"]=\"\"; using Pkg; Pkg.add(\"Conda\"); using Conda; Conda.add(\"matplotlib\"); Pkg.add(\"PyCall\"); Pkg.build(\"PyCall\"); Pkg.add(\"PyPlot\");'"
-
-codecov: true
-
-matrix:
-  allow_failures:
-    - julia: nightly
-  include:
-    - stage: "Documentation"
-      julia: 1.5
-      os: linux
-      script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
-        - julia --project=docs/ docs/make.jl
-      after_success: skip
-
-after_success: 
-  - "julia -e 'import Pkg; Pkg.add(\"Coverage\"); using Coverage; Coveralls.submit(Coveralls.process_folder())'"
-
