### 2022-02-17 99e1d5f
Adds doc build to ci.yml
