### 2022-02-17 5d4653e
fixes mass function
