+++ .github/workflows/ci.yml
@@ -60,8 +60,8 @@ jobs:
       - run: |
           julia --project=docs -e '
             using Documenter: doctest
-            using MYPACKAGE
-            doctest(MYPACKAGE)' # change MYPACKAGE to the name of your package
+            using ProbabilityBoundsAnalysis
+            doctest(ProbabilityBoundsAnalysis)' # change MYPACKAGE to the name of your package
       - run: julia --project=docs docs/make.jl
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
