### 2022-02-18 ca1ec73
nearly working doctests
