+++ docs/Project.toml
@@ -1,5 +1,6 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
 ProbabilityBoundsAnalysis = "cd73f0b2-6f06-11ea-242f-83a09de6a7cf"
 
 [compact]
