+++ .github/workflows/ci.yml
@@ -59,7 +59,7 @@ jobs:
             Pkg.instantiate()'
       - run: |
           julia --project=docs -e '
-            using Documenter: doctest
+            using Documenter
             using ProbabilityBoundsAnalysis
             DocMeta.setdocmeta!(ProbabilityBoundsAnalysis, :DocTestSetup, :(using ProbabilityBoundsAnalysis, IntervalArithmetic, Random; Random.seed!(31415)); recursive=true)
             doctest(ProbabilityBoundsAnalysis)' # change MYPACKAGE to the name of your package
