### 2022-02-18 526a6e6
working doctests & uses doctests for testing
