+++ test/runtests.jl
@@ -11,5 +11,5 @@ include("pbox/pbox.jl")
 include("pbox/distributions.jl")
 include("pbox/arithmetic.jl")
 
-DocMeta.setdocmeta!(ProbabilityBoundsAnalysis, :DocTestSetup, :(using ProbabilityBoundsAnalysis, IntervalArithmetic, Random; Random.seed!(31415)); recursive=true)
+DocMeta.setdocmeta!(ProbabilityBoundsAnalysis, :DocTestSetup, :(using ProbabilityBoundsAnalysis, IntervalArithmetic, Random; Random.seed!(MersenneTwister(), 31415)); recursive=true)
 doctest(ProbabilityBoundsAnalysis, manual = false)
