### 2022-02-18 95d16e6
fixes doctest
