+++ src/pbox/distributions.jl
@@ -759,11 +759,11 @@ julia> a = N(0,1);
 
 julia> rand(a,5)
 5-element Vector{Interval{Float64}}:
-  [0.38532, 0.398856]
- [-0.772194, -0.755415]
-  [1.28155, 1.31058]
- [-0.189119, -0.176374]
- [-0.553385, -0.538836]
+  [0.538836, 0.553385]
+ [-0.331854, -0.318639]
+  [0.125661, 0.138305]
+ [-1.40508, -1.3722]
+ [-0.31864, -0.30548]
 ```
 See also: [`cut`](@ref), [`cdf`](@ref), [`mass`](@ref)
 """
