+++ docs/make.jl
@@ -3,7 +3,7 @@
 
 using ProbabilityBoundsAnalysis, IntervalArithmetic, Documenter
 
-DocMeta.setdocmeta!(ProbabilityBoundsAnalysis, :DocTestSetup, :(using ProbabilityBoundsAnalysis, IntervalArithmetic, Random; Random.seed!(MersenneTwister(), 31415)); recursive=true)
+DocMeta.setdocmeta!(ProbabilityBoundsAnalysis, :DocTestSetup, :(using ProbabilityBoundsAnalysis, IntervalArithmetic, Random; Random.seed!(31415)); recursive=true)
 
 makedocs(
     modules = [ProbabilityBoundsAnalysis],
