+++ test/runtests.jl
@@ -10,6 +10,3 @@ using Test, ProbabilityBoundsAnalysis, IntervalArithmetic, Distributions, Docume
 include("pbox/pbox.jl")
 include("pbox/distributions.jl")
 include("pbox/arithmetic.jl")
-
-DocMeta.setdocmeta!(ProbabilityBoundsAnalysis, :DocTestSetup, :(using ProbabilityBoundsAnalysis, IntervalArithmetic, Random; Random.seed!(31415)); recursive=true)
-doctest(ProbabilityBoundsAnalysis, manual = false)
