+++ .github/workflows/ci.yml
@@ -12,6 +12,7 @@ jobs:
         version:
           - '1.5'
           - '1.6'
+          - 'nightly'
           
         os:
           - ubuntu-latest
