+++ src/pbox/dependency.jl
@@ -1,693 +1,0 @@
-######
-# This file is part of the ProbabilityBoundsAnalysis.jl package.
-#
-#   Definition of bivariate copula class, joint distribution class and methods
-#
-#           University of Liverpool, Institute for Risk and Uncertainty
-#
-#                                           Author: Ander Gray
-#                                           Email:  ander.gray@liverpool.ac.uk
-######
-
-###
-#   Notes: This is a script for performing tests with copulas for plotting, sampling and
-#   finding resulting copulas from arithemtic between uncertain variables.
-#
-#   Would it be possible to represent the set of all copulas as an expansion of orthogonal
-#   polynomials? Could be used as a sentivitivty analysis for dependancy. Something equivilant
-#   to disperive Monte Carlo, where not only the correlatin coefficient is changed but the entire
-#   dependancy structure. Could this expansion be used in the arithmetic?
-#
-#   Or can we do it as sums of the Frechet bounds? (Shuffles of M)
-#   See: Nelson page 68.
-#
-#   For algebra of random variables, including algebra of covaraince See https://en.wikipedia.org/wiki/Algebra_of_random_variables
-#       -> The above link is also useful for moment propagation. Could you perform covariance algebra when the covarainces are partially known
-#
-#
-#   One model for dependency trancking, is to assume a copula family, and compute the covariance of the results of opartaions with their inputs
-#   using the arithemtic of covaraince described above. For this we would need an interval covaraince algebra, and a way of performing convolutions with it.
-#
-###
-
-###
-#
-#   Known bugs:
-#                   ->  The differentiation for the sampling should not depend on n. May request values x>1 or x<0  (Solved)
-#                   ->  eps() should be used for differentiation. Liams code. (Solved)
-#                   ->  Linear interpolator not working... (Solved)
-#                   ->  Bilinear interpolartor? (Solved)
-#
-###
-
-using Distributions
-#using PyPlot
-using PyCall
-using LinearAlgebra
-using Interpolations
-#using3D()
-include("NormalDistribution.jl")
-
-#global  n  = 200;        # Number of descretizations
-global  pn = 200;        # Number of descretization for plotting
-#global bOt = 0.001;              # smallest quamtile to use if left tail is unbounded
-#global tOp = 0.999;              # largest quamtile to use if right tail is unbounded
-
-abstract type AbstractCopula <: Real end
-abstract type AbstractJoint <: Real end
-
-mutable struct copula <: AbstractCopula
-
-    cdf ::  Array{Float64,2}
-    density :: Union{Array{Float64,2},Missing}
-    func :: Union{Function,Missing}
-    param :: Union{Float64,Missing}
-
-    function copula(cdf = missing; density = missing, func = missing, param = missing)
-
[…]
