+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"copulas/#Modelling-dependencies-with-copulas","page":"Dependence","title":"Modelling dependencies with copulas","text":"","category":"section"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"Any possible multivariate dependence can be encoded in a copula. Copulas, sometimes called dependency functions, are a joint cdf with standard uniform marginals, and are a way to model dependency independently of marginal distributions. ","category":"page"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"The supported copulas are:","category":"page"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"π: independence\nM: perfect dependence and the upper Frechét bound\nW: opposite dependence and the lower Frechét bound\nGaussian: family with correlation coefficient r (-1 for W, 0 for π, 1 for M)\nFrank: family with parameter s (0 for W, 1 for π, Inf for M)\nClayton: family with parameter t (-1 for W, 0 for π, Inf for M)","category":"page"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"A copula can be created and plotted in the following way:","category":"page"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"julia> C = GauCopula(0.7)\n Copula ~ Gau(r=0.7)\n\njulia> plot(C)","category":"page"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"(Image: \"Gaussian copula with a correlation of 0.7\")","category":"page"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"A copula C is the function C01^d rightarrow 01, where d is the dimension of the copula. Only bivariate copulas are considered so far in ProbabilityBoundsAnalysis.jl. A copula can be evaluated and sampled in the following way:","category":"page"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"julia> C(0.5,0.5)\n [0.370898, 0.375924]\n\njulia> C([0.2,0.6],[0.3,0.7])\n2×2 Array{Interval{Float64},2}:\n [0.140274, 0.144236]  [0.191086, 0.196042]\n [0.270231, 0.275108]  [0.524632, 0.53028]\n \njulia> samps = sample(C,10^4);","category":"page"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"Note that like distributions and p-boxes, in ProbabilityBoundsAnalysis.jl samples and evaluations of copulas are intervals. This also allows copulas to be imprecisely defined (set of dependencies). An imprecise copula can be created by giving the constructor and interval:","category":"page"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"julia> C = GauCopula(interval(-0.5,0.5))\n Imp Copula ~ Gaussian(r = [-0.5, 0.5])\n\njulia> C(0.5,0.5)\n [0.164165, 0.33585]\n\njulia> C([0.2,0.6],[0.3,0.7])\n2×2 Array{Interval{Float64},2}:\n [0.0162205, 0.116083]  [0.082187, 0.18411]\n [0.105729, 0.248291]   [0.350841, 0.495683]\n\njulia> plot(a)","category":"page"},{"location":"copulas/","page":"Dependence","title":"Dependence","text":"(Image: \"Imprecise Gaussian copula with a
[…]
