### 2022-02-21 af9781a
removes globals and adds constant struct for parameters
