### 2022-03-04 #96 @AnderGray
@dpsanders @baggepinnen Thanks both for this review, and for the great suggestions which will improve the quality of package substantially. Apologies I'm taking so long to get back. I'm working on two PRs which address your comments. I've completed David's, and will hopefully finish Fredrik's in the coming week
