### 2022-03-04 AnderGray/ProbabilityBoundsAnalysis.jl#32 @AnderGray
plot recipes is the ideal way to do this. However when I tried the plotting with Plots.jl, I couldn't quite get it to look the way I liked, so I'd prefer to stick with PyPlot for now. What I have done is remove PyPlot as a dependence, and used Requires.jl. So that if you wish to plot, you need to `using PyPlot`

I still need to make note of this in the Documentation.
