### 2022-03-04 #96 @AnderGray
> 👋 @AnderGray: It's not clear to me why Eduardo Patelli is an author of the paper, since he does not seem to be an author of the package, according to the docs.

Edoardo Patelli provided funding sourcing and mentorship, without which this package would have not been created, and I believe he should be an author of the paper. I will change the authorship in the Docs
