### 2022-03-08 aac27a2
fixes arithmetic between p-boxes and intervals
