### 2022-03-08 108884c
adds initial comparisons between scalars, intervals and p-box
