### 2022-03-08 1696bc0
adds tests for comparisons
