+++ test/runtests.jl
@@ -10,4 +10,5 @@ using Test, ProbabilityBoundsAnalysis, IntervalArithmetic, Distributions, Docume
 include("pbox/pbox.jl")
 include("pbox/distributions.jl")
 include("pbox/arithmetic.jl")
+include("pbox/comparisons.jl")
 include("pbox/parameter_change.jl")
