### 2022-03-08 e5eac96
adds initial copula switching
