### 2022-03-08 bb03891
adds more comparisons and comparison tests
