### 2022-03-08 0af809e
revert arithm between intervals and p-boxes
