### 2022-03-08 518b530
adds testing for plots
