+++ test/runtests.jl
@@ -12,3 +12,4 @@ include("pbox/distributions.jl")
 include("pbox/arithmetic.jl")
 include("pbox/comparisons.jl")
 include("pbox/parameter_change.jl")
+include("pbox/plots.jl")
