### 2022-03-08 1c11061
simplify plot tests
