+++ test/pbox/plots.jl
@@ -16,22 +16,37 @@
 
         using PyPlot
 
-        @test @isdefined plot(a, plotting = false)
 
-        @test @isdefined plot(a, name = "same", plotting = false)
-        @test @isdefined plot(a, col = "red", plotting = false)
-        @test @isdefined plot(a, alpha = 0.5, plotting = false)
-        @test @isdefined plot(a, fontsize = 20, plotting = false)
-        @test @isdefined plot(a, false, plotting = false)
+
+        pl1 = plot(a, plotting = false)
+        pl2 = plot(a, name = "same", plotting = false)
+        pl3 = plot(a, col = "red", plotting = false)
+        pl4 = plot(a, alpha = 0.5, plotting = false)
+        pl5 = plot(a, fontsize = 20, plotting = false)
+        pl6 = plot(a, false, plotting = false)
+
+        @test @isdefined pl1
+        @test @isdefined pl2
+        @test @isdefined pl3
+        @test @isdefined pl4
+        @test @isdefined pl5
+        @test @isdefined pl6
 
         b = interval(2, 3)
 
-        @test @isdefined plot(b, plotting = false)
-        @test @isdefined plot(b, name = "same", plotting = false)
-        @test @isdefined plot(b, col = "red", plotting = false)
-        @test @isdefined plot(b, alpha = 0.5, plotting = false)
-        @test @isdefined plot(b, fontsize = 20, plotting = false)
-        @test @isdefined plot(b, false, plotting = false)
+        pl1_i = plot(b, plotting = false)
+        pl2_i = plot(b, name = "same", plotting = false)
+        pl3_i = plot(b, col = "red", plotting = false)
+        pl4_i = plot(b, alpha = 0.5, plotting = false)
+        pl5_i = plot(b, fontsize = 20, plotting = false)
+        pl6_i = plot(b, false, plotting = false)
+
+        @test @isdefined pl1_i
+        @test @isdefined pl2_i
+        @test @isdefined pl3_i
+        @test @isdefined pl4_i
+        @test @isdefined pl5_i
+        @test @isdefined pl6_i
 
     end
 
@@ -41,20 +56,29 @@
 
         PyPlot.ioff()
 
-        @test @isdefined plot(a)
-        @test @isdefined plot(a, name = "same")
-        @test @isdefined plot(a, fontsize = 18, alpha = 0.7)
-        @test @isdefined plotStep(a)
+        pl1 = plot(a)
+        pl2 = plot(a, name = "same")
+        pl3 = plot(a, fontsize = 18, alpha = 0.7)
+        pl4 = plotStep(a)
+
+        @test @isdefined pl1
+        @test @isdefined pl2
+        @test @isdefined pl3
+        @test @isdefined pl4
 
         b = N(0,1)
 
         j1 = bivpbox(b, b, a);
 
-        @test @isdefined plot(j1)
+        plj_1 = plot(j1)
+
+        @test @isdefined plj_1
 
         samps = ProbabilityBoundsAnalysis.sample(j1, 10)
 
-        @test @isdefined scatter(samps)
+        plj_2 = scatter(samps)
+
+        @test @isdefined plj_2
     end
 
     PyPlot.ion()
