+++ test/pbox/plots.jl
@@ -75,10 +75,15 @@
         @test @isdefined plj_1
 
         samps = ProbabilityBoundsAnalysis.sample(j1, 10)
-
         plj_2 = scatter(samps)
 
         @test @isdefined plj_2
+
+        samps_precise = rand(100)
+        plj_3  = scatter(samps_precise)
+
+        @test @isdefined plj_3
+
     end
 
     PyPlot.ion()
