### 2022-03-21 63380a6
adds test for p-box from interval data
