### 2022-03-21 ef970f0
fix plot tests and adds more mass tests
