### 2022-03-21 306f726
exports \subset and fixes tests and function
