### 2022-03-21 7b4d2f2
adds tests for special functions
