+++ test/runtests.jl
@@ -8,8 +8,9 @@
 using Test, ProbabilityBoundsAnalysis, IntervalArithmetic, Distributions, Documenter
 
 include("pbox/pbox.jl")
+include("pbox/special.jl")
+include("pbox/comparisons.jl")
 include("pbox/distributions.jl")
 include("pbox/arithmetic.jl")
-include("pbox/comparisons.jl")
 include("pbox/parameter_change.jl")
 include("pbox/plots.jl")
