### 2022-03-23 e56a017
adds tests for intersection and union of p-boxes
