+++ src/pbox/distributions.jl
@@ -55,7 +55,7 @@ function env(x...; naRm = false )
         u = min.(u,elts[i].u);
         d = max.(d,elts[i].d);
         ml = min(ml,elts[i].ml);
-        mh = max(mh,elts[i].ml);
+        mh = max(mh,elts[i].mh);
         vl = min(vl,elts[i].vl);
         vh = max(vh,elts[i].vh);
         bounded[1] = min(bounded[1], elts[i].bounded[1]);
