### 2022-03-30 cbaf976
add parametric distribution tests
