+++ test/pbox/distributions.jl
@@ -9,7 +9,35 @@ using Distributions
 
 @testset "Distributions" begin
 
-    Nsamples = 5000; Ncdf = 5000;
+    Nsamples = 1000; Ncdf = 1000;
+
+    function test_dist(x :: pbox, a)
+
+        samps = rand(Nsamples);
+
+        @test length(x.d) == ProbabilityBoundsAnalysis.parametersPBA.steps
+
+        pbaSamples = cut.(x,samps);
+        distSamps  = quantile.(a,samps);
+
+        @test all( distSamps .∈ pbaSamples )
+
+        testVals = range(left(x), stop = right(x),length = Ncdf)
+        cdfPba = cdf.(x, testVals);
+        cdfDis = cdf.(a, testVals);
+
+        @test all( cdfDis .∈ cdfPba)
+
+        EndPoints = rand(Nsamples,2) .* (right(x) .- left(x)) .+ left(x);
+        EndPoints = sort(EndPoints, dims = 2);
+
+        massInts = interval.(EndPoints[:,1], EndPoints[:,2])
+        massPba = mass.(x,massInts)
+        massDis = cdf.(a,EndPoints[:,2]) .- cdf.(a, EndPoints[:,1])
+
+        @test all( massDis .∈ massPba)
+
+    end
 
     @testset "Set based operations" begin
 
@@ -55,113 +83,128 @@ using Distributions
 
     end
 
-    @testset "Normal" begin
+    @testset "Parametric p-boxes - 200 steps" begin
 
-        x = normal(interval(0,1),interval(1,2));
+        test_dist(normal(0,1), Normal(0,1))
+        test_dist(normal(0..1, 1..2), Normal(0,1))
 
-        @test mean(x) == interval(0,1);
-        @test var(x) == interval(1,2)^2;
-        @test x.shape == "normal"
-        @test all( map(!,x.bounded))
+        test_dist(uniform(0, 1), Uniform(0,1))
+        test_dist(uniform(0..0.5, 1..1), Uniform(0,1))
 
-        x = normal(0,1);
-        a = Normal(0,1);
+        test_dist(beta(2,3), Beta(2, 3))
+        test_dist(beta(2..3, 3..4), Beta(2, 3))
 
-        samps = rand(Nsamples);
+        test_dist(betaPrime(2, 3), BetaPrime(2, 3))
+        test_dist(betaPrime(2..3, 3..4), BetaPrime(2, 3))
 
-        pbaSamples = cut.(x,samps);
-        distSamps  = quantile.(a,samps);
+        test_dist(biweight(1,2), Biweight(1, 2))
+        test_dist(biweight(1..2,2..2), Biweight(1, 2))
 
-        @test all( distSamps .∈ pbaSamples )
+        test_dist(cauchy(0, 1), Cauchy(0, 1))
+        test_dist(cauchy(0..0.5, 1), Cauchy(0, 1))
 
-        testVals = range(left(x), stop = right(x),length = Ncdf)
-        cdfPba = cdf.(x, testVals);
-        cdfDis = cdf.(a, testVals);
+        test_dist(chi(1), Chi(1))
+        test_dist(chi(1..2), Chi(1))
 
-        @test all( cdfDis .∈ cdfPba)
+        test_dist(chisq(1), Chisq(1))
+        test_dist(chisq(1..2), Chisq(1))
 
-        EndPoints = rand(Nsamples,2) .* (right(x) .- left(x)) .+ left(x);
-        EndPoints = sort(EndPoints, dims = 2);
[…]
