### 2022-03-30 a86f6fa
removes old code and tests uniform from mean and std
