### 2022-03-30 5ee2791
adds == for p-boxes
