### 2022-03-30 547e220
exports == and adds tests more tests for distribution free
