### 2022-03-30 92f9723
adds initial copula tests
