+++ test/runtests.jl
@@ -11,6 +11,7 @@ include("pbox/pbox.jl")
 include("pbox/special.jl")
 include("pbox/comparisons.jl")
 include("pbox/distributions.jl")
+include("pbox/copulas.jl")
 include("pbox/arithmetic.jl")
 include("pbox/parameter_change.jl")
 include("pbox/plots.jl")
