### 2022-03-31 #96 @carstenbauer
What's the status here? @AnderGray I see there are still open issues and PRs related to the reviews by @dpsanders and @baggepinnen. Am I right in assuming that you just need more time to improve / fix things?
