### 2022-03-31 #96 @AnderGray
@carstenbauer Thanks for following this up. Yes, most of the issues have been addressed. @dpsanders's questions have all be addressed I believe, and although the PR was merged (since we wanted this functionality). I've addressed some of @baggepinnen's, the biggest is related to testing, which is taking time. Nearly there
