### 2022-03-31 51b6ef1
adds conditional and massU, massD copula tests
