+++ test/pbox/copulas.jl
@@ -5,56 +5,62 @@
 #
 ######
 
+using ProbabilityBoundsAnalysis: Frechet
+
 Ngrid = 1000;
 πfunc(x,y) = x*y
 mfunc(x,y) = min(x, y)
 wfunc(x,y) = max(x + y - 1, 0)
 
+function test_copula(C, func)
+
+    xs = range(0, 1, length = Ngrid)
+
+    @test func(0.4, 0.4) ∈ cdf(C,0.4,0.4)
+    @test func(0.4, 0.4) ∈ C(0.4,0.4)
+
+    cdfs = cdf(C, xs, xs)
+    cdfs_lo = cdfD(C, xs, xs)
+    cdfs_hi = cdfU(C, xs, xs)
+
+    real_vals = [func(x, y) for x in xs, y in xs]
+
+    @test all(real_vals .∈ cdfs)
+    @test all(real_vals .>= getfield.(cdfs_lo, :lo))
+    @test all(real_vals .<=  getfield.(cdfs_hi, :hi))
+
+    @test all(func.(0.4, xs) .∈ C(0.4,xs))
+    @test all(func.(xs, 0.4) .∈ C(xs, 0.4))
+
+    @test all(func.(0.4, xs) .∈ C(0.4,collect(xs)))
+    @test all(func.(collect(xs), 0.4) .∈ C(xs, 0.4))
+
+    @test all(func.(xs, xs) .∈ C.(xs, xs))
+
+    x_ints = mince(0..1, 100)
+
+    @test all(range(0,1, length = 100) .∈ C(1, x_ints))
+    @test all(range(0,1, length = 100) .∈ C(x_ints, 1))
+
+    @test all(zeros(100) .∈ C(0, x_ints))
+    @test all(zeros(100) .∈ C(x_ints, 0))
+
+    cdf_ints = [C(x, y) for x in x_ints, y in x_ints]
+    cdf_int_real = [func(x, y) for x in x_ints, y in x_ints]
+
+    @test all(cdf_int_real .⊆ cdf_ints)
+
+    cop_masses = [mass(C, interval(0, x), interval(0, y)) for x in xs, y in xs]
+
+    cop_massesU = [massU(C, interval(0, x), interval(0, y)) for x in xs, y in xs]
+    cop_massesD = [massD(C, interval(0, x), interval(0, y)) for x in xs, y in xs]
+
+    @test all(real_vals .∈ cop_masses)
+    @test all(real_vals .∈ cop_massesU)
+    @test all(real_vals .∈ cop_massesD)
+end
+
 @testset "Copulas" begin
-
-    function test_copula(C, func)
-
-        xs = range(0, 1, length = Ngrid)
-
-        @test func(0.4, 0.4) ∈ cdf(C,0.4,0.4)
-        @test func(0.4, 0.4) ∈ C(0.4,0.4)
-
-        cdfs = cdf(C, xs, xs)
-        cdfs_lo = cdfD(C, xs, xs)
-        cdfs_hi = cdfU(C, xs, xs)
-
-        real_vals = [func(x, y) for x in xs, y in xs]
-
-        @test all(real_vals .∈ cdfs)
-        @test all(real_vals .>= getfield.(cdfs_lo, :lo))
-        @test all(real_vals .<=  getfield.(cdfs_hi, :hi))
-
-        @test all(func.(0.4, xs) .∈ C(0.4,xs))
-        @test all(func.(xs, 0.4) .∈ C(xs, 0.4))
-
-        @test all(func.(0.4, xs) .∈ C(0.4,collect(xs)))
-        @test all(func.(collect(xs), 0.4) .∈ C(xs, 0.4))
-
[…]
