### 2022-03-31 b2fd102
adds show tests
