+++ test/runtests.jl
@@ -14,4 +14,5 @@ include("pbox/distributions.jl")
 include("pbox/copulas.jl")
 include("pbox/arithmetic.jl")
 include("pbox/parameter_change.jl")
+include("pbox/show.jl")
 include("pbox/plots.jl")
