### 2022-04-02 18183eb
fixes frank parameter for Inf and -Inf
