+++ src/pbox/copulas.jl
@@ -404,7 +404,7 @@ Pi() = πCop()
 
 function Frank(s = 1)                       #   s is real; inf for perfect, 0 for indep, -inf for opposite
 
-    if !(s ∈ interval(-Inf,Inf)); throw(ArgumentError("coefficient must be ∈ [-Inf, Inf]\n $s ∉ [-Inf, Inf]"));end
+    if !(s ∈ interval(-Inf,Inf)) && s != -Inf && s != Inf; throw(ArgumentError("coefficient must be ∈ [-Inf, Inf]\n $s ∉ [-Inf, Inf]"));end
 
     if s == -Inf         # Limit should be set earlier
         C = W()
@@ -428,7 +428,7 @@ end
 
 function Frank(s :: Interval)
 
-    if !(s ⊆ interval(-Inf,Inf)) && s != -Inf && s != Inf; throw(ArgumentError("coefficient must be ⊆ [-Inf, Inf]\n $s ⊄ [-Inf, Inf]"));end
+    if !(s ⊆ interval(-Inf,Inf)); throw(ArgumentError("coefficient must be ⊆ [-Inf, Inf]\n $s ⊄ [-Inf, Inf]"));end
 
     lb = Frank(s.lo);
     ub = Frank(s.hi);
