+++ test/runtests.jl
@@ -12,6 +12,7 @@ include("pbox/special.jl")
 include("pbox/comparisons.jl")
 include("pbox/distributions.jl")
 include("pbox/copulas.jl")
+include("pbox/bivariate_pbox.jl")
 include("pbox/arithmetic.jl")
 include("pbox/parameter_change.jl")
 include("pbox/show.jl")
