+++ test/pbox/bivariate_pbox.jl
@@ -111,5 +111,4 @@ end
         test_2pbox(j4, func7)
 
     end
-
 end
