### 2022-04-03 43a9c3c
removes redundant moment transformation code
