### 2022-04-03 28da4c8
removes old arithmetic code and adds further frechet tests
