+++ src/pbox/arithmetic.jl
@@ -119,8 +119,6 @@ function convPerfect(x::pbox, y::pbox; op = +)
         cd = map(op, x.d[:], y.d[:]);
     end
 
-    #cu = map(op, x.u[:],y.u[:]);
-    #cd = map(op, x.d[:],y.d[:]);
     scu = sort(cu);
     scd = sort(cd);
 
@@ -145,8 +143,6 @@ function convOpposite(x::pbox, y::pbox; op = +)
         cd = map(op, x.d[:], y.d[end:-1:1]);
     end
 
-    #cu = map(op, x.u[:], y.u[end:-1:1]);
-    #cd = map(op, x.d[:], y.d[end:-1:1]);
     scu = sort(cu);
     scd = sort(cd);
 
@@ -193,12 +189,6 @@ function convFrechet(x::pbox, y::pbox; op = +)
     vl = 0;
     vh = Inf;
 
-    if (op)∈([+,-])                 # Was commented below, can include onve mom prop is finished
-        #zv = env(xv+yv-2* sqrt(xv*yv), xv+yv+2* sqrt(xv*yv))
-        # vh <- x@v+y@v - 2*sqrt(x@v*y@v)
-        # vl <- x@v+y@v + 2*sqrt(x@v*y@v)
-    end
-
     bounded = min(x.bounded,y.bounded);
     return pbox(zu, zd, ml = ml, mh = mh, vl = vl, vh = vh, bounded=bounded);
 end
@@ -212,9 +202,6 @@ function sigma(x::pbox, y ::pbox; op = +,  C = πCop()::AbstractCopula)
 
     #if (op == -) return sigma(x, negate(y), op=+, C = C); end
     #if (op == /) return sigma(x, reciprocate(y), op = *, C = C);end
-
-    #x = makepbox(x);
-    #y = makepbox(y);
 
     xus = x.u; xds = x.d;
     yus = y.u; yds = y.d;
@@ -323,102 +310,12 @@ function calcSigmaBoundRight(x, y , op,  C, bounded)
 
     bounded ? is =  ProbabilityBoundsAnalysis.jj() : is = ProbabilityBoundsAnalysis.jjj()
 
-
     for i = 2:Ns    # Cannot be threaded, because of 2nd line
         these   = findall(is[i-1] .<= cdf .<= is[i]);
         !isempty(these) ? z[i-1] = maximum(zs[these]) : (i != 2 ?  z[i-1] = z[i-2] : continue)
     end
 
     return z
-
-end
-
-
-
-function sigmaOld(x::pbox, y ::pbox; op = +,  C = πCop()::AbstractCopula)
-
-    #if (op == -) return sigma(x, negate(y), op=+, C = C); end
-    #if (op == /) return sigma(x, reciprocate(y), op = *, C = C);end
-
-    x = makepbox(x);
-    y = makepbox(y);
-
-    Ns = parametersPBA.steps
-
-    zus = [map(op, ux, uy) for ux in x.u[1:end-1], uy in y.u[1:end-1]]        # Carteesian products
-    zds = [map(op, dx, dy) for dx in x.d[2:end], dy in y.d[2:end]]
-
-    zd = zeros(Ns);
-    zu = zeros(Ns);
-
-    zd[1] = minimum(zds); zd[end] = maximum(zds);
-    zu[1] = minimum(zus); zu[end] = maximum(zus);
-
-    zus = zus[:];
-    zds = zds[:];
-
-    uMasses = C.cdfU[2:end, 2:end] + C.cdfU[1:end-1, 1:end-1] - C.cdfU[1:end-1, 2:end] - C.cdfU[2:end, 1:end-1]
[…]
