### 2022-04-03 904b42e
adds tests for partially defined dependence
