### 2022-04-06 4450b86
adds simple unary tests
