+++ paper/ref.bib
@@ -1,14 +1,24 @@
 @article{bezanson2017julia,
-  title={Julia: A fresh approach to numerical computing},
-  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
-  journal={SIAM review},
-  volume={59},
-  number={1},
-  pages={65--98},
-  year={2017},
-  publisher={SIAM}
-}
+author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+title = {Julia: A Fresh Approach to Numerical Computing},
+journal = {SIAM Review},
+volume = {59},
+number = {1},
+pages = {65-98},
+year = {2017},
+doi = {10.1137/141000671},
 
+URL = { 
+        https://doi.org/10.1137/141000671
+    
+},
+eprint = { 
+        https://doi.org/10.1137/141000671
+    
+}
+,
+    abstract = { Bridging cultures that have often been distant, Julia combines expertise from the diverse fields of computer science and computational science to create a new approach to numerical computing. Julia is designed to be easy and fast and questions notions generally held to be “laws of nature" by practitioners of numerical computing: \beginlist \item High-level dynamic programs have to be slow. \item One must prototype in one language and then rewrite in another language for speed or deployment. \item There are parts of a system appropriate for the programmer, and other parts that are best left untouched as they have been built by the experts. \endlist We introduce the Julia programming language and its design---a dance between specialization and abstraction. Specialization allows for custom treatment. Multiple dispatch, a technique from computer science, picks the right algorithm for the right circumstance. Abstraction, which is what good computation is really about, recognizes what remains the same after differences are stripped away. Abstractions in mathematics are captured as code through another technique from computer science, generic programming. Julia shows that one can achieve machine performance without sacrificing human convenience. }
+}
 @article{Moore1990,
 	author = {Moore, James T. and Bard, Jonathan F.},
 	doi = {10.1287/opre.38.5.911},
@@ -33,10 +43,17 @@
   publisher={Prentice-Hall Englewood Cliffs, NJ}
 }
 @book{jaulin2001interval,
-  title={Applied Interval analysis},
-  author={Jaulin, Luc and Kieffer, Michel and Didrit, Olivier and Walter, Eric},
-  year={2001},
-  publisher={Springer}
+  TITLE = {{Applied Interval Analysis with Examples in Parameter and State Estimation, Robust Control and Robotics}},
+  AUTHOR = {Jaulin, Luc and Kieffer, Michel and Didrit, Olivier and Walter, Eric},
+  URL = {https://hal.archives-ouvertes.fr/hal-00845131},
+  NOTE = {http://www.springer.com/engineering/computational+intelligence+and+complexity/book/978-1-4471-1067-5},
+  PUBLISHER = {{Springer London Ltd}},
+  PAGES = {398},
+  YEAR = {2001},
+  MONTH = Aug,
+  PDF = {https://hal.archives-ouvertes.fr/hal-00845131/file/contentaia_LISA2001.pdf},
+  HAL_ID = {hal-00845131},
+  HAL_VERSION = {v1},
 }
 @book{feller1968probability,
   title={An Introduction to Probability Theory and Its Applications Volume 1.},
@@ -91,26 +108,37 @@
   title={Statistical reasoning with imprecise probabilities},
   author={Walley, Peter},
   year={1991},
-  publisher={Chapman and Hall, London}
+  publisher={Chapman and Hall, London},
+  isbn = {9780412286605}
 }
 @book{troffaes2014lower,
-  title={Lower previsions},
-  author={Troffaes, Matthias C.M. and De Cooman, Gert},
-  year={2014},
-  publisher={John Wiley \& Sons}
[…]
