### 2022-04-13 3209aae
simplifies pbox intersection
