### 2022-04-13 e639b31
adds documentation for set based operations
