+++ docs/src/sets.md
@@ -1,10 +1,122 @@
 # Comparisons and Set based operations
 
-Comparisons between p-boxes, intervals, and scalars
+Comparisons
 ---
 
-![\Large x=\frac{-b\pm\sqrt{b^2-4ac}}{2a}](https://latex.codecogs.com/svg.latex?x%3D%5Cfrac%7B-b%5Cpm%5Csqrt%7Bb%5E2-4ac%7D%7D%7B2a%7D)
+Comparisons `(<, >, <=, >=, ==)` between p-boxes, intervals and scalars can be performed. However unlike for comparisons between real numbers which yield Boolean values (`true` or `false`), comparisons with p-boxes generally yield interval probabilities, giving the uncertainty that the random variable characterised by a p-box meets the condition.
 
+__*NOTE: unlike `IntervalArithmetic.jl`, comparisons in `ProbabilityBoundsAnalysis.jl` will generally give non-Boolean values (interval probabilities). This may cause crashes when evaluating control-flow (if-else) with uncertainty*__
+
+### Comparisons of p-boxes and scalars
+
+For a p-box `X` and real number `y`, `X <= y` is the evaluation of y in the CDF of X:
+
+![](https://latex.codecogs.com/png.image?%5Cinline%20%5Ctiny%20%5Cdpi%7B300%7DX%20%5Cleq%20y%20%5Cto%20F_%7BX%7D(y))
+
+
+Similarly, `X >= y` is
+
+![compare1](https://latex.codecogs.com/png.image?%5Cinline%20%5Ctiny%20%5Cdpi%7B300%7DX%20%5Cgeq%20y%20%5Cto%201%20-%20F_%7BX%7D(y))
+
+Example
+```julia
+julia> X = uniform(0, 1)
+julia> X <= 0.7
+[0.695, 0.705001]
+
+julia> X > 0.4
+[0.594999, 0.605]
+
+julia> X = normal(interval(-0.5, 0.5), interval(1, 1.5))
+julia> X >= 1
+[0.0649999, 0.37]
+```
+
+Boolean values are returned if condition is gauranteed
+
+```julia
+julia> X = uniform(0, 1)
+julia> X <= 2
+true
+
+julia> X > 2
+false
+```
+
+### Comparisons of p-boxes and intervals
+
+A comparison between a p-box `X` and an interval `Y` can be evaluated as follows
+
+![](https://latex.codecogs.com/png.image?%5Cinline%20%5Ctiny%20%5Cdpi%7B300%7DX%20%5Cleq%20Y%20%5Cto%20X%20-%20Y%20%5Cleq%200)
+
+
+where subtraction is evaluated with p-box arithmetic, and then the resulting p-box's CDF is evaluated at `0`. I.e., `Z = X - Y` and then `cdf(Z, 0.0)`. Similarly
+
+![ass](https://latex.codecogs.com/png.image?%5Cinline%20%5Ctiny%20%5Cdpi%7B300%7DX%20%5Cgeq%20Y%20%5Cto%20Y%20-%20X%20%5Cleq%200)
+
+Example
+```julia
+julia> X = uniform(0, 1)
+julia> Y = interval(0.7, 2)
+julia> X <= Y
+[0.695, 1]
+
+julia> X > Y
+[0, 0.305]
+
+julia> X <= interval(2, 3)
+true
+
+julia> X > interval(2, 3)
+false
+```
+
+### Comparisons between p-boxes
+
+Comparison between two p-boxes `X` and `Y` are performed similarly to intervals
+
+![](https://latex.codecogs.com/png.image?%5Cinline%20%5Ctiny%20%5Cdpi%7B300%7DX%20%5Cleq%20Y%20%5Cto%20X%20-%20Y%20%5Cleq%200)
+
[…]
