### 2022-04-14 0e6b421
fix punctuation and removes footnotes
