### 2022-04-14 138f6a0
further removes footnotes
