+++ paper/paper.tex
@@ -21,18 +21,6 @@ An arithmetic of probability distributions has long held interest among mathemat
 The goal of \texttt{ProbabilityBoundsAnalysis.jl} is to \textbf{compute guaranteed bounds on functions of random variables given only partial knowledge of the input probability distributions and their dependencies}. That is, to compute with partial knowledge about the input joint distribution. Ideally, all of the available information about random variables should be used, but without making any extra assumptions beyond what are justified by what is known.
 
 The idea of bounding probability has a long tradition throughout the history of probability theory. George Boole~\cite{boole1854investigation, hailperin1986boole} used the notion of interval bounds on probability. Chebyshev~\cite{chebyshev1874valeurs} described bounds on a distribution when only the mean and variance of the variable are known, and Markov~\cite{markoff1900question} found bounds on a positive variable when only the mean is known.  Fréchet~\cite{frechet1935generalisation} discovered how to bound joint distributions solely from knowing the marginal distributions, without making independence assumptions. Bounding probabilities has continued to the present day, culminating into the modern theory of imprecise probabilities~\cite{walley1991statistical, klir2013uncertainty, troffaes2014lower, augustin2014introduction}.
-
-Imprecise probability is effectively a generalisation of probability theory where uncertainty can be expressed about the probability measure. This is particularly relevant when information is scarce, unreliable, vague, conflicting or imprecise. In such cases defining a unique probability distribution is difficult. P-boxes are one of many ways to describe a set of distributions. Others include: Dempster-Shafer structures~\cite{dempster2008upper,shafer1976mathematical}, random sets~\cite{molchanov2005theory}, possibility distributions~\cite{zadeh1978fuzzy,dubois1988possibility, hose2019possibilistic}, lower previsions~\cite{troffaes2014lower}, and credal sets~\cite{levi1983enterprise}. These structures were discovered independently, but are often synonymous and can be translated from one to another, with different degrees of generality. Imprecise probability unifies the underlying theories. For a comprehensive overview of the theory, and a formal description of uncertainty and information in terms of these structures, \cite{klir2013uncertainty} is recommended.
-
-
-
-\iffalse
-Probability bounds analysis is a combination of the methods of standard interval analysis \cite{moore1996interval, jaulin2001interval} and classical probability theory (see, inter alia, \cite{feller1968probability, feller1971probability}).  The idea of bounding probability has a very long tradition throughout the history of probability theory. Indeed, George Boole \cite{boole1854investigation, hailperin1986boole} used the notion of interval bounds on probability. Chebyshev \cite{chebyshev1874valeurs} described bounds on a distribution when only the mean and variance of the variable are known, and Markov \cite{markoff1900question} found bounds on a positive variable when only the mean is known. Fréchet \cite{frechet1935generalisation} discovered how to make calculations with uncertain estimates of joint probabilities without making independence assumptions. Bounding probabilities has continued to the present day \cite{walley1991statistical, klir2013uncertainty, troffaes2014lower, augustin2014introduction}, culminating into the modern theory of Imprecise Probabilities.
-
-An arithmetic of probability distributions has held a long interest among mathematicians and engineers
-
-Computations with probabilities are usually performed by Monte-Carlo style simulations, where essentially many random realisations of functions are required to be run. These sampling methods require many thousands of realisations to be accurate, and even then will only produce an approximation of the desired probabilistic quantity. In contrast, the methods ProbabilityBoundsAnalysis.jl are exact rather than approximate, and give no restriction to the distribution shape or dependency.
-\fi
 \begin{figure*}[htp]
 
[…]
