+++ paper/paper.tex
@@ -133,18 +133,18 @@ P-boxes generalise precise distributions and intervals in the following way. A d
 \end{align*}
 \noindent where $\epsilon_k$ is
 \begin{equation*}
-   \epsilon_k(x) = \begin{cases} 0 &\text{when } x < k \\ 1 &\text{when } x \geq k. \end{cases} 
+   \epsilon_k(x) = \begin{cases} 0 &\text{when } x < k, \\ 1 &\text{when } x \geq k. \end{cases} 
 \end{equation*}
 \noindent Moreover, theoretical bounds on the mean and variance of an interval can be found~\cite{ferson2002ramas}
 \begin{align*}
   \mu &\in [a, b], \\
   \sigma^2 &\in [0, (b - a)^{2}/4].
 \end{align*}
-\noindent That is, it is not possible to find a distribution inside $[a, b]$ with a variance greater than $(b-a)^{2}/4$. The lower bound on the variance is zero, since scalars are also included in the interval. Under this definition of an interval, a random sample will always be the interval $[a, b]$, the cdf returns 0 when $x < a$, the interval $[0,1]$ when $a \leq x < b$, and $1$ when $b \leq x$. Further, the probability measure of the interval $X = [a, b]$ will be: 
+\noindent That is, it is not possible to find a distribution inside $[a, b]$ with a variance greater than $(b-a)^{2}/4$. The lower bound on the variance is zero, since scalars are also included in the interval. Under this definition of an interval, a random sample will always be the interval $[a, b]$, the cdf returns 0 when $x < a$, the interval $[0,1]$ when $a \leq x < b$, and $1$ when $b \leq x$. Further, the probability measure of the interval $X = [a, b]$ is
 \begin{equation*}
   \mathbb{P}_{X}(U) = \begin{cases}
-    0 & \text{when } U \cap [a,b] = \emptyset \\
-    1 & \text{when } [a,b] \subseteq U  \\
+    0 & \text{when } U \cap [a,b] = \emptyset, \\
+    1 & \text{when } [a,b] \subseteq U,  \\
     [0, 1] & \text{otherwise.}\\
   \end{cases}
 \end{equation*}
@@ -267,19 +267,19 @@ It is well known that generally a random variables distribution function cannot 
 
 These p-boxes are constructed using the Chebyshev (mean and variance)~\cite{chebyshev1874valeurs}, Markov (mean and one bound)~\cite{markoff1900question} and Cantelli (mean and range) inequalities from probability theory. These inequalities allow for the cdf to be bounded from various range and moment information. For example, the Chebyshev inequality can be used to find p-box bounds from the random variables mean and variance
 \begin{equation*}
-  \underline{F}_{X}(x) = \begin{cases} 1 / (1 + (x - \mu_{X})^2 / \sigma^{2}_{X}),& \text{if } \mu_{X}<x \\
-    1,& \text{if } x\leq \mu_{X} \end{cases}
+  \underline{F}_{X}(x) = \begin{cases} 1 / (1 + (x - \mu_{X})^2 / \sigma^{2}_{X}) & \text{if } \mu_{X}<x, \\
+    1 & \text{if } x\leq \mu_{X}, \end{cases}
 \end{equation*}
 \begin{equation*}
-  \overline{F}_{X}(x)  = \begin{cases} 1 / (1 + \sigma^{2}_{X}/(x - \mu_{X})^2),& \text{if } x<\mu_{X} \\
[…]
