### 2022-04-14 79e2d2b
refactor introduction text
