+++ paper/paper.tex
@@ -5,9 +5,6 @@
 \usepackage{amsmath}
 \usepackage{xcolor}
 %\bibliographystyle{juliacon}
-\usepackage[labelformat=simple]{subfig}
-\renewcommand\thesubfigure{\arabic{subfigure}}
-\renewcommand{\subfigurename}{Figure}
 \begin{document}
 \input{header}
 \maketitle
