### 2022-04-15 #96 @AnderGray
@baggepinnen @dpsanders apologies for the wait, I think we're ready for another look. **Unsure why why the pdf isn't compiling**, but I believe I've incorporated the requests. I've added all the DOI's I could find, and shortened the paper.

For @baggepinnen, the PR can be seen in [pull/37](https://github.com/AnderGray/ProbabilityBoundsAnalysis.jl/pull/37):
- Testing improved
- Documentation improved (included added community guidelines)
- Added comparisons

Note that unlike intervals, which always return Boolean values (perhaps erroneously), comparisons between p-boxes return interval probabilities, which give the probability that the imprecisely characterise random variables meet the condition. More details can be found in the [new docs](https://github.com/AnderGray/ProbabilityBoundsAnalysis.jl/blob/baggepinnen_review/docs/src/sets.md)

Example between p-box and scalars:
```julia
julia> X = uniform(0, 1)
julia> X <= 0.7
[0.695, 0.705001]

julia> X >= 0.4
[0.594999, 0.605]

julia> X = normal(interval(-0.5, 0.5), interval(1, 1.5))
julia> X >= 1
[0.0649999, 0.37]

julia> X = uniform(0, 1)
julia> X <= 2
true

julia> X >= 2
false
```

Example between p-box and intervals
```julia
julia> X = uniform(0, 1)
julia> Y = interval(0.7, 2)
julia> X <= Y
[0.695, 1]

julia> X >= Y
[0, 0.305]

julia> X <= interval(2, 3)
true

julia> X >= interval(2, 3)
false
```
Example between p-boxes
```julia
julia> X = uniform(0, 1)
julia> Y = uniform(0.5, 1.5)
julia> X <= Y
[0.5, 1]

julia> X <= uniform(2, 3)
true

julia> X >= uniform(2, 3)
false
```
Note that this will cause your `foo` function above to fail due to using a non-Boolean ... I do however believe that this is correct in this context. Evaluate this with intervals `[0, 1] <= 0.5` returns `false` by default, but it could be argued that the interval Boolean `[0, 1]` is the correct return here, since we cannot say if the interval is above or below the scalar. `ProbabilityBoundsAnalysis.jl` returns the probability that the condition is met, e.g., `uniform(0,1) <= 0.5 ` gives a tight probability interval `[0.494999, 0.505001]` (the interval comes from discretisation error bounding).

A warning about this has been added to the docs. **Ideally I would like to modify the Julia warning message when an interval (probability) is used in a Boolean context, saying why this is happening. Do you know if it's possible to do this?**

For @dpsanders the PR was already merged, but is [pull/35](https://github.com/AnderGray/ProbabilityBoundsAnalysis.jl/pull/35). I've also responded to your issues individually, which are still open.
