### 2022-04-15 #96 @dpsanders
There is a massive overhaul of IntervalArithmetic.jl underway which will fix the `X < Y` issue for intervals `X` and `Y`, by allowing the user to specify the behaviour they prefer: https://github.com/JuliaIntervals/IntervalArithmetic.jl/pull/271

@AnderGray : Thanks, I'll take a look!
