### 2022-04-19 47c2258
adds further typeographic improvements
