+++ paper/paper.tex
@@ -29,9 +29,9 @@ The idea of bounding probability has a long tradition throughout the history of 
   \label{fig:figure1}
   
 \end{figure*}
-Imprecise probability is effectively a generalisation of probability theory where uncertainty can be expressed about the probability measure. This is particularly relevant when information is scarce, unreliable, vague, conflicting, or imprecise: cases where defining a unique probability distribution is difficult. P-boxes are one of many ways to describe a set of distributions. Others include: Dempster-Shafer structures~\cite{dempster2008upper,shafer1976mathematical}, random sets~\cite{molchanov2005theory}, possibility distributions~\cite{zadeh1978fuzzy,dubois1988possibility, hose2019possibilistic}, lower previsions~\cite{troffaes2014lower}, and credal sets~\cite{levi1983enterprise}. These structures were discovered independently, but are often synonymous (with different degrees of generality) and can be transformed from one to another. Imprecise probability unifies the underlying theories. For an overview of the theory, and a formal description of uncertainty and information in terms of these structures, work by George Klir~\cite{klir2013uncertainty} is recommended.
+Imprecise probability is effectively a generalisation of probability theory where uncertainty can be expressed about the probability measure. This is particularly relevant when information is scarce, unreliable, vague, conflicting, or imprecise: cases where defining a unique probability distribution is difficult. P-boxes are one of many ways to describe a set of distributions. Others include: Dempster-Shafer structures~\cite{dempster2008upper,shafer1976mathematical}, random sets~\cite{molchanov2005theory}, possibility distributions~\cite{dubois1988possibility, hose2019possibilistic}, lower previsions~\cite{troffaes2014lower}, and credal sets~\cite{levi1983enterprise}. These structures were discovered independently, but are often synonymous (with different degrees of generality) and can be transformed from one to another. Imprecise probability unifies the underlying theories. For an overview of the theory, and a formal description of uncertainty and information in terms of these structures, work by George Klir~\cite{klir2013uncertainty} is recommended.
 \subsection{Automatic uncertainty in a programming language: why Julia?}
-A potential long term goal of probability bounds analysis is to create a uncertain programming language, where any computer variable may be represented as an interval, distribution, p-box, or other imprecise quantity. Such a framework would allow for uncertain extensions of deterministic functions to be computed in an automatic, rigorous, and tight fashion. Those with little knowledge of probability theory could add uncertainty to their computations with minimal effort. Stochastic computation often require specialist training to be used effectively, and are usually not rigorous and are famously computationally expensive. \texttt{ProbabilityBoundsAnalysis.jl} provides the arithmetic basis for such a framework. In this section we give some arguments why Julia~\cite{bezanson2017julia} is an ideal target language for such a framework.
[…]
