### 2022-04-19 100dfc1
updates arithmetic docs
