+++ docs/src/arithmetic.md
@@ -28,41 +28,70 @@ Supported dependent arithmetic between uncertain numbers:
 Most of the fundamental binary operations can be performed between uncertain numbers of all types:
 
 ```julia
-julia> a = normal(-1,1); 
-julia> b = interval(1,2);
+julia> a = normal(10,1); 
+julia> b = uniform(1,2);
 julia> a + b
- Pbox: 	  ~  ( range=[-3.09023,4.090232], mean=[0.0,1.0], var=[1.0,1.25])
+ Pbox: 	  ~  ( range=[7.9098, 15.09], mean=11.5, var=[0.38392, 2.1086])
 
 julia> a - b
- Pbox: 	  ~  ( range=[-6.090232,1.0902323], mean=[-3.0,-2.0], var=[1.0,1.25])
+ Pbox: 	  ~  ( range=[4.9098, 12.09], mean=8.5, var=[0.38392, 2.1086])
 
 julia> a * b
- Pbox: 	  ~  ( range=[-4.090232,4.180464], mean=[-1.015451,-1.9690], var=[0.99763,3.99053])
+ Pbox: 	  ~  ( range=[6.9098, 26.18], mean=[12.194, 18.069], var=[0.6576, 32.544])
 
 julia> a / b
- Pbox: 	  ~  ( range=[-2.045116,2.0902323], mean=[-0.50772,-0.984548], var=[0.249408,0.99763])
+ Pbox: 	  ~  ( range=[3.4549, 13.09], mean=[5.6205, 8.3861], var=[0.13574, 7.6644])
 ```
-
-All of the above operations assume independence. For unknown dependence:
+```@raw html
+<img src="./plots/sumFrechet.png" width="49.5%"/>
+<img src="./plots/subFrechet.png" width="49.5%"/>
+<img src="./plots/prodFrechet.png"width="49.5%"/>
+<img src="./plots/divFrechet.png" width="49.5%"/>
+```
+By default all arithmetic operations in `ProbabilityBoundsAnalysis.jl` are _Frechet_ operations, operations which consider all possible dependencies (copulas). Therefore we naturally get p-boxes even if we began with precise distributions. For independence:
 ```julia
-julia> convFrechet(a, b, op = +)
- Pbox: 	  ~  ( range=[-3.09023,4.090232], mean=[0.0,1.0], var=[0.383917,2.1086384])
+julia> convIndep(a, b, op = +)
+ Pbox: 	  ~  ( range=[7.9098, 15.09], mean=11.5, var=1.0833)
 
-julia> convFrechet(a, b, op = -)
- Pbox: 	  ~  ( range=[-6.09023,1.090232], mean=[-3.0,-2.0], var=[0.383917,2.108638])
+julia> convIndep(a, b, op = -)
+ Pbox: 	  ~  ( range=[4.9098, 12.09], mean=8.5, var=1.0833)
+
+julia> convIndep(a, b, op = *)
+ Pbox: 	  ~  ( range=[6.9098, 26.18], mean=15.0, var=[10.052, 11.538])
+
+julia> convIndep(a, b, op = /)
+ Pbox: 	  ~  ( range=[3.4549, 13.09], mean=[6.919, 6.944], var=[2.3032, 2.6815])
 ```
 
-The resulting p-boxes are much wider than the independence case.
+The resulting p-boxes are precise.
+
+```@raw html
+<img src="./plots/sumIndep.png" width="49.5%"/>
+<img src="./plots/subIndep.png" width="49.5%"/>
+<img src="./plots/prodIndep.png"width="49.5%"/>
+<img src="./plots/divIndep.png" width="49.5%"/>
+```
 
[…]
