### 2022-04-19 6224da7
updates arithmetic documentation
