### 2022-05-08 f84cdcd
adds automatic drawing
