### 2022-05-14 205e763
simplifies linear interpolator
