+++ test/pbox/pbox.jl
@@ -35,7 +35,7 @@
 
     @test !isinterval(x)
     @test x.ml == x.mh
-    @test x.ml == 0.5
+    @test 0.5 ∈ mean(x)
     @test x.vh == x.vl
 
     x = pbox(interval(0,4), ml=1, mh = 3, vl = 0.2, vh = 2)
