### 2022-05-14 cc2ec44
simplifies interpolator in constructor
