+++ test/pbox/pbox.jl
@@ -34,9 +34,7 @@
     x = pbox([0,1]);
 
     @test !isinterval(x)
-    @test x.ml == x.mh
     @test 0.5 ∈ mean(x)
-    @test x.vh == x.vl
 
     x = pbox(interval(0,4), ml=1, mh = 3, vl = 0.2, vh = 2)
 
