### 2022-12-28 #96 @carstenbauer
@AnderGray I apologise sincerely and must admit that I completely forgot about your submission / this issue. Are you willing to continue and move this over the finish line?

If so, it would be great if you could debug the PDF compilation error (Does it work locally? Maybe "rebase" the `paper` folder to the current version of the submission template.). Afterwards, I'll ask the reviewers to give this another / a final look.

Sorry again.
